IGI/LR5/src/hotelapp/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-phonenumber-field==7.3.0
phonenumbers==8.13.37
pillow==10.3.0
sqlparse==0.5.0
